fix: disable banner if is __SERVER__ and handle diffView
