+++ src/helpers/GDPRCookies.js
@@ -37,7 +37,6 @@ class GDPRCookies {
   }
 
   set(name, value) {
-    console.log(name, value, this.cookie_name);
     if (!this.cookie_name) {
       return;
     }
